Finally rendering

+++ src/util/VectorUtilities.cpp
@@ -5,12 +5,12 @@ Vector3f VectorUtilities::crossProduct(const Vector3f &a, const Vector3f &b)
 
     return Vector3f(a.y * b.z - a.z * b.y, a.z * b.x - a.x * b.z, a.x * b.y - a.y * b.x);
 }
-float VectorUtilities::dotProduct(const Vector3f &a, const Vector3f &b)
+double VectorUtilities::dotProduct(const Vector3f &a, const Vector3f &b)
 {
-    float scalar = 0;
+    double scalar = 0;
     for (int i = 0; i < 3; i++)
     {
-        scalar += a.vec[i] * b.vec[i];
+        scalar += (double)a.vec[i] * (double)b.vec[i];
     }
 
     return scalar;
@@ -26,18 +26,18 @@ Vector3f VectorUtilities::normalizeCopy(const Vector3f &v)
     vec.normalize();
     return vec;
 }
-float VectorUtilities::Length(const Vector3f &a, const Vector3f &b)
+double VectorUtilities::Length(const Vector3f &a, const Vector3f &b)
 {
     return sqrt(VectorUtilities::dotProduct(a, b));
 }
 //Vector 4f
 
-float VectorUtilities::dotProduct(const Vector4f &a, const Vector4f &b)
+double VectorUtilities::dotProduct(const Vector4f &a, const Vector4f &b)
 {
-    float scalar = 0;
+    double scalar = 0;
     for (int i = 0; i < 4; i++)
     {
-        scalar += a.vec[i] * b.vec[i];
+        scalar += (double)a.vec[i] * b.vec[i];
     }
 
     return scalar;
@@ -54,7 +54,7 @@ Vector4f VectorUtilities::normalizeCopy(const Vector4f &v)
     return vec;
 }
 
-float VectorUtilities::Length(const Vector4f &a, const Vector4f &b)
+double VectorUtilities::Length(const Vector4f &a, const Vector4f &b)
 {
     return sqrt(VectorUtilities::dotProduct(a, b));
 }
